requirements.txt
python-pptx==0.6.21
google-generativeai==0.3.2
google-search-results==2.4.2
python-dotenv==1.0.0
